replication moved under cluster actor
