+++ src/init.rs
@@ -1,5 +1,5 @@
 use crate::env_var;
-use crate::services::cluster::replications::replication::{ReplicationInfo, IS_MASTER_MODE};
+use crate::services::cluster::actors::replication::{ReplicationInfo, IS_MASTER_MODE};
 
 pub struct Environment {
     pub replicaof: Option<(String, String)>,
