+++ src/presentation/cluster_in/outbound/stream.rs
@@ -3,9 +3,9 @@ use crate::presentation::cluster_in::connection_manager::ClusterConnectionManage
 use crate::presentation::cluster_in::create_peer;
 use crate::services::cluster::actors::commands::AddPeer;
 use crate::services::cluster::actors::commands::ClusterCommand;
+use crate::services::cluster::actors::replication::ReplicationInfo;
 use crate::services::cluster::peers::identifier::PeerIdentifier;
 use crate::services::cluster::peers::kind::PeerKind;
-use crate::services::cluster::replications::replication::ReplicationInfo;
 use crate::services::interface::TRead;
 use crate::services::interface::TWrite;
 use crate::services::statefuls::snapshot::snapshot_applier::SnapshotApplier;
