+++ src/services/cluster/actors/commands.rs
@@ -2,8 +2,8 @@ use crate::services::aof::{WriteOperation, WriteRequest};
 
 use crate::services::cluster::peers::identifier::PeerIdentifier;
 
+use crate::services::cluster::actors::replication::{HeartBeatMessage, ReplicationInfo};
 use crate::services::cluster::peers::peer::Peer;
-use crate::services::cluster::replications::replication::{HeartBeatMessage, ReplicationInfo};
 
 pub enum ClusterCommand {
     AddPeer(AddPeer),
