+++ src/services/cluster/actors/mod.rs
@@ -1,3 +1,5 @@
 pub(crate) mod actor;
 pub mod commands;
 mod consensus;
+pub(crate) mod replication;
+pub mod replid_generator;
