+++ src/services/cluster/mod.rs
@@ -1,4 +1,3 @@
 pub(crate) mod actors;
 pub mod listeners;
 pub(crate) mod peers;
-pub(crate) mod replications;
