--- src/services/cluster/replications/mod.rs
@@ -1,2 +1,0 @@
-pub(crate) mod replication;
-pub(crate) mod replid_generator;
